Performance update & Renaming
+ Sender can now be closed! (closes socket)
+ GraphiteConnector closes Sender before shutdown
+ NOW: Sender could be null for GraphiteConnector -> Sender is now set in method "run(Sender sender)" instead of setting it in the constructor
+ SocketManager now in package "server.transfer.sender.connection"
