+++ src/main/java/server/transfer/sender/ConsoleSender.java
@@ -26,4 +26,9 @@ public class ConsoleSender extends Sender {
 		System.out.println(payload.asString());
 	}
 
+	@Override
+	public void close() {
+		
+	}
+
 }
